fix missing call to snmp_store_needed()
The writeable scalars of system_mib and snmp_mib dit not trigger
the snmp_store to save the value after commit.
The netsnmp_watcher_helper_handler() do now call snmp_store_needed()
after commit. This may be not always needed,
but at least snmp_mib and system_mib needs this!

The netsnmp_inject_handler() is not longer used, it prevents setting
of snmp_enableauthentrapsset, which is needed to save persistence value.